+++ src/Admin/SettingsPage.php
@@ -124,8 +124,26 @@ class SettingsPage extends \WC_Settings_Page {
 
             [ 'type' => 'kis_system_audit' ],
 
+            [
+                'title' => __( 'OP Kassa Environment', 'woocommerce-kis' ),
+                'desc'  => __( 'Check the "Connect to OP Kassa Test Environment"-option below to use Kassa Test Environment and save the settings. <div class="env-change-warning">WARNING: Changing the environment disconnects Your WooCommerce instance from Kassa, if connected.</div>', 'woocommerce-kis' ),
+                'type'  => 'title',
+                'id'    => 'kis_environment_settings',
+            ],
+
+            'kis_test_environment_enabled' => [
+                'type'    => 'checkbox',
+                'title'   => __( 'Connect to OP Kassa Test Environment', 'woocommerce-kis' ),
+                'id'      => 'kis_test_environment_enabled',
+            ],
+
+            [
+                'type' => 'sectionend',
+                'id'   => 'kis_environment_settings',
+            ],
+
             [ 'type' => 'kis_oauth' ],
-            
+
             [
                 'type' => 'sectionend',
                 'id'   => 'kis_oauth_settings',
@@ -233,6 +251,11 @@ class SettingsPage extends \WC_Settings_Page {
         $oauth           = new OAuth();
         $oauth_activated = $oauth->is_oauth_active();
 
+        $kis_test_environment_enabled = KIS_WOOCOMMERCE_TEST_ENVIRONMENT_ENABLED === "yes" ? "test" : "production";
+        $kis_test_environment_enabled = KIS_HAS_CUSTOM_ENVIRONMENT ? 
+            'CUSTOM (environmental urls probably set up in wp-config.php. These settings override the environment selection.)' : 
+            $kis_test_environment_enabled;
+
         $oauth_url        = $this->get_kassa_oauth_url();
         $cancel_oauth_url = $this->get_oauth_cancel_url();
         ?>
@@ -243,8 +266,13 @@ class SettingsPage extends \WC_Settings_Page {
                             <?php $this->the_merchant_details(); ?>
                             <div class="postbox">
                                 <h2><span><?php esc_html_e( 'Integration status', 'woocommerce-kis' ); // phpcs:ignore ?></span></h2>
-                                <div class="inside">
+                                <div class="inside">                           
                                     <div class="kis-oauth">
+                                        <div class="kis-environment">
+                                            <input type="hidden" name="kis_has_custom_environment" value="<?= KIS_HAS_CUSTOM_ENVIRONMENT; ?>" />
+                                            <span class="kis-environment-title"><?= esc_html_e( 'OP Kassa Environment', 'woocommerce-kis' ) . ': '; ?></span>
+                                            <span class="kis-environment-value"><?= esc_html_e( $kis_test_environment_enabled, 'woocommerce-kis' ); ?></span>
+                                        </div>
                                         <div class="kis-oauth__section kis-oauth__section--kassa">
                                             <?php if ( $oauth_activated ) : ?>
                                                 <p>