+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2020-09-11
+
+### Changed
+- Added possibility to select the target Kassa environment (test/production).
+- Kassa connection is now disconnected on plugin deactivation.
+- Fixed calls to wp_timezone(), which threw an error on older (< WP 5.3) WordPress versions.
+
 ## [0.7.11] - 2020-07-02
 
 ### Changed